requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
groq>=0.9.0
tavily-python>=0.3.3
chromadb>=0.5.0
reportlab>=4.2.2
python-dotenv>=1.0.0
pydantic>=2.8.0
httpx>=0.27.0
websockets>=12.0
